+++ packages/nodes-base/nodes/Calendly/v1/GenericFunctions.ts
@@ -10,7 +10,7 @@ import type {
 
 function getAuthenticationTypeFromApiKey(data: string): 'accessToken' | 'apiKey' {
 	// The access token is a JWT, so it will always include dots to separate
-	// header, payoload and signature.
+	// header, payload and signature.
 	return data.includes('.') ? 'accessToken' : 'apiKey';
 }
